% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcPPM}
\alias{calcPPM}
\title{calcPPM()}
\usage{
calcPPM(mz, ppm = 20)
}
\arguments{
\item{mz}{Target \emph{m/z}}

\item{ppm}{Tolerance parameter}
}
\value{
numeric list
}
\description{
Calculates the bounds around a mass for a given ppm threshold.
}
\examples{
calcPPM(198.1000, 20)
}
