% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convertDtoMZML}
\alias{convertDtoMZML}
\title{convertDtoMZML()}
\usage{
convertDtoMZML(file, path.out = paste0(file, "/_mzML"), MRM = FALSE)
}
\arguments{
\item{file}{.d data file to convert}

\item{path.out}{Output directory for converted files. Defaults to 'path/_mzmL'}
}
\description{
Convert Agilent .d file to .mzML. Can only be MS1 scan data.
}
\examples{
convertDtoMZML(system.file('extdata', '_d/0001.d', package='LCMStools'))
}
