% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCEChroms}
\alias{getCEChroms}
\title{Get chromatograms for each collision energy in experiment}
\usage{
getCEChroms(chromatograms)
}
\arguments{
\item{chromatograms}{A list of chromatogram objects}
}
\value{
A long data frame of retention times, intensities, collision
energies, and data file
}
\description{
Returns a data frame of retention times and intensities (chromatograms) along
with the associated collision energy of the MS experiment
}
\examples{
library(MSnbase)
ex.file = system.file('extdata', '_d/0176.d/0176.mzML', package='LCMStools')
ex.data = readSRMData(ex.file)
getCEChroms(ex.data)
}
