% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMStools-package.R
\docType{package}
\name{LCMStools-package}
\alias{LCMStools}
\alias{LCMStools-package}
\title{LCMStools: Frequently Used Functions for LC/MS Data}
\description{
Tools for LC/MS data analysis. Data conversion requires msconvert installed from ProteoWizard.
}
\author{
\strong{Maintainer}: Josh Morimoto \email{jmorim@mit.edu} (\href{https://orcid.org/0000-0001-5309-6796}{ORCID})

}
\keyword{internal}
