% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseMSMethod}
\alias{parseMSMethod}
\title{parseMSMethod()}
\usage{
parseMSMethod(xml.file)
}
\arguments{
\item{xml.file}{The 192_1.xml MS method file in the data file's method folder}
}
\value{
Nested tibble of method parameters, including scan segments, source
parameters, and chromatograms
}
\description{
Parses whole MS method using above parsers. Combine with map_dfr to build a
table for multiple method files
}
\examples{
parseMSMethod(
 system.file('extdata', '_d/0001.d/AcqData/opt_dmrm.m/192_1.xml', package='LCMStools')
)
library(purrr)
system.file('extdata', '_d', package='LCMStools') |> 
 dir(pattern='192_1.xml', recursive=TRUE, full.names=TRUE) |> 
 map_df(parseMSMethod)
}
