% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseSampleInfo}
\alias{parseSampleInfo}
\title{parseSampleInfo()}
\usage{
parseSampleInfo(sample.info.xml)
}
\arguments{
\item{sample.info.xml}{Path of the sample_info.xml file in the data file directory}
}
\value{
named list
}
\description{
Parses a data file's sample_info.xml to a named vector
Combine with map_dfr to generate a data table for multiple data files
}
\examples{
parseSampleInfo(
 system.file('extdata', '_d/0001.d/AcqData/sample_info.xml', package='LCMStools')
)
}
